ui: refactor inbox view into separate components
- Rename `InboxDecryptMessageView` to `DecryptMessageView` and move to separate file
- Rename `MessagesPresentInbox` to `MessagesList` to move to own component file
- Move `NoMessagesInbox` to components directory
- Renamed `InboxView` to the correct folder
