+++ Localizable.xcstrings
@@ -1756,6 +1756,9 @@
         }
       }
     },
+    "Decrypt Message" : {
+
+    },
     "default_app_scheme" : {
       "extractionState" : "manual",
       "localizations" : {
